add custom ip list to cmd

+++ client/testnet.go
@@ -9,6 +9,7 @@ import (
 	"net"
 	"os"
 	"path/filepath"
+	"strings"
 
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 	"github.com/spf13/cobra"
@@ -48,8 +49,8 @@ var (
 	flagNodeDaemonHome    = "node-daemon-home"
 	flagNodeCLIHome       = "node-cli-home"
 	flagStartingIPAddress = "starting-ip-address"
-	flagCoinDenom         = "coin-denom"
 	flagKeyAlgo           = "algo"
+	flagIpAddrs           = "ip-addrs"
 )
 
 const nodeDirPerm = 0755
@@ -78,13 +79,14 @@ Note, strict routability for addresses is turned off in the config file.`,
 			nodeDaemonHome, _ := cmd.Flags().GetString(flagNodeDaemonHome)
 			nodeCLIHome, _ := cmd.Flags().GetString(flagNodeCLIHome)
 			startingIPAddress, _ := cmd.Flags().GetString(flagStartingIPAddress)
+			ipAddresses, _ := cmd.Flags().GetString(flagIpAddrs)
 			numValidators, _ := cmd.Flags().GetInt(flagNumValidators)
 			coinDenom, _ := cmd.Flags().GetString(flagCoinDenom)
 			algo, _ := cmd.Flags().GetString(flagKeyAlgo)
 
 			return InitTestnet(
 				cmd, config, cdc, mbm, genAccIterator, outputDir, chainID, coinDenom, minGasPrices,
-				nodeDirPrefix, nodeDaemonHome, nodeCLIHome, startingIPAddress, keyringBackend, algo, numValidators,
+				nodeDirPrefix, nodeDaemonHome, nodeCLIHome, startingIPAddress, ipAddresses, keyringBackend, algo, numValidators,
 			)
 		},
 	}
@@ -95,6 +97,7 @@ Note, strict routability for addresses is turned off in the config file.`,
 	cmd.Flags().String(flagNodeDaemonHome, "ethermintd", "Home directory of the node's daemon configuration")
 	cmd.Flags().String(flagNodeCLIHome, "ethermintcli", "Home directory of the node's cli configuration")
 	cmd.Flags().String(flagStartingIPAddress, "192.168.0.1", "Starting IP address (192.168.0.1 results in persistent peers list ID0@192.168.0.1:46656, ID1@192.168.0.2:46656, ...)")
+	cmd.Flags().String(flagIpAddrs, "", "List of IP addresses to use (i.e. `192.168.0.1,172.168.0.1` results in persistent peers list ID0@192.168.0.1:46656, ID1@172.168.0.1)")
 	cmd.Flags().String(flags.FlagChainID, "", "genesis file chain-id, if left blank will be randomly created")
 	cmd.Flags().String(flagCoinDenom, ethermint.AttoPhoton, "Coin denomination used for staking, governance, mint, crisis and evm parameters")
 	cmd.Flags().String(server.FlagMinGasPrices, fmt.Sprintf("0.000006%s", ethermint.AttoPhoton), "Minimum gas prices to accept for transactions; All fees in a tx must meet this minimum (e.g. 0.01aphoton,0.001stake)")
@@ -118,6 +121,7 @@ func InitTestnet(
 	nodeDaemonHome,
 	nodeCLIHome,
 	startingIPAddress,
+	ipAddresses,
 	keyringBackend,
 	algo string,
 	numValidators int,
@@ -133,6 +137,13 @@ func InitTestnet(
 
 	if err := sdk.ValidateDenom(coinDenom); err != nil {
 		return err
+	}
+
+	var ips []string
+	if ipAddresses != "" {
+		ipAddresses := strings.Replace(ipAddresses, " ", "", -1)
+		ips = strings.Split(ipAddresses, ",")
+		numValidators = len(ips)
 	}
 
 	nodeIDs := make([]string, numValidators)
@@ -169,10 +180,16 @@ func InitTestnet(
 
 		config.Moniker = nodeDirName
 
-		ip, err := getIP(i, startingIPAddress)
-		if err != nil {
-			_ = os.RemoveAll(outputDir)
-			return err
+		var ip string
+		var err error
+		if ipAddresses == "" {
+			ip, err = getIP(i, startingIPAddress)
+			if err != nil {
+				_ = os.RemoveAll(outputDir)
+				return err
+			}
+		} else {
+			ip = ips[i]
 		}
 
 		nodeIDs[i], valPubKeys[i], err = genutil.InitializeNodeValidatorFiles(config)
